Use HTTP 507 Error to indicate that local blobstore disk is full
We extended CloudFoundry API errors with 500000: NoSpaceOnDevice.
When local blobstore is used and the filesystem where it is located
is full, HTTP 507 will be returned.

This does not affect how nginx processes uploads. When nginx encounters
problems with its temporary storage, which may be in the same filesystem,
it will return a different error before request reaches bits-service.

[#115615679]

Signed-off-by: Alexander Egurnov <egurnov@de.ibm.com>
Signed-off-by: Norman Sutorius <norman.sutorius@de.ibm.com>