Implement package status updates for CloudController
Bits-service is now able to report package status to CloudController through an
encrypted channel with client key checking (Mutual TLS).

It is still possible to run bits-service standalone when CCUpdater config part
is missing. The new config subtree is called `bits-service.cc_updates`.

When CC updates are enabled, special setup within CC is performed to allow
the intergration tests to run properly. When CCUpdater is not configured, the
setup is skipped.

[#143968055] [#143967581]

Signed-off-by: Alexander Egurnov <egurnov@de.ibm.com>
Signed-off-by: Steffen Uhlig <Steffen.Uhlig@de.ibm.com>
Signed-off-by: Norman Sutorius <norman.sutorius@de.ibm.com>