Add blobstore job only to WebDAV deployments
This avoids a failing deploy if there is no cf release on a director
that gets the bits-service deployed. The blobstore job is part of cf and
not available standalone.

As we expect the WebDAV backend to go away once the bits-service is the
default blobstore, I decided that it is not worth trying to deploy
bits-service with WebDAV, but without cf.

[#148279417]