Remove release_spec.rb
This test is not useful anymore, because we are testing this in upload_limit_spec.rb. Also, with our current deployment and test structure it is not possible to get to the non-blobstore-specific max_body_size property. The only reason this worked with nginx was because nginx checked the body-size before even routing the request to the correct endpoint.

[#156533708]